requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
bleach==4.1.0
branca==0.4.2
category-encoders==2.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
chromedriver-autoinstaller==0.2.2
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
cycler==0.10.0
dateparser==1.0.0
Django==3.2.6
djangorestframework==3.12.4
djangotoolbox==1.8.0
djongo==1.3.6
et-xmlfile==1.1.0
filelock==3.0.12
idna==3.2
importlib-metadata==4.6.3
iniconfig==1.1.1
Jinja2==3.0.1
joblib==1.0.1
jsonlines==2.0.0
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.2
munch==2.5.0
nltk==3.6.2
numpy==1.21.1
openpyxl==3.0.9
packaging==21.0
pandas==1.3.1
pandas-stubs==1.2.0.30
patsy==0.5.1
plotly==5.1.0
pluggy==0.13.1
protobuf==3.17.3
psycopg2==2.7.7
psycopg2-binary==2.9.1
py==1.10.0
pydot==1.4.2
pymongo==3.12.0
pyparsing==2.4.7
pyproj==3.1.0
pytest==6.2.4
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.28
requests==2.26.0
scikit-learn==0.24.2
scipy==1.7.1
Shapely==1.7.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.2.4
statsmodels==0.12.2
tenacity==8.0.1
threadpoolctl==2.2.0
toml==0.10.2
tqdm==4.62.0
typing-extensions==3.10.0.0
tzlocal==3.0
uritools==3.0.2
urlextract==1.3.0
urllib3==1.26.6
webencodings==0.5.1
zipp==3.5.0
